+++ addons/calendar/static/src/activity/activity_service_patch.js
@@ -4,15 +4,6 @@ import { ActivityService } from "@mail/core/web/activity_service";
 import { patch } from "@web/core/utils/patch";
 
 patch(ActivityService.prototype, {
-    insert(data) {
-        const activity = super.insert(...arguments);
-        const { calendar_event_id: calendarEventId } = data;
-        if (calendarEventId) {
-            activity["calendar_event_id"] = calendarEventId;
-        }
-        return activity;
-    },
-
     async rescheduleMeeting(activityId) {
         const action = await this.orm.call("mail.activity", "action_create_calendar_event", [
             [activityId],
