+++ addons/im_livechat/static/src/embed/core/message_model_patch.js
@@ -1,9 +1,21 @@
 /* @odoo-module */
 
+import { ChatbotStep } from "@im_livechat/embed/chatbot/chatbot_step_model";
+
 import { Message } from "@mail/core/common/message_model";
 
 import { patch } from "@web/core/utils/patch";
 import { session } from "@web/session";
+
+patch(Message, {
+    insert(data) {
+        const message = super.insert(data);
+        if (data.chatbotStep) {
+            message.chatbotStep = new ChatbotStep(data.chatbotStep);
+        }
+        return message;
+    },
+});
 
 patch(Message.prototype, {
     get isSelfAuthored() {
