+++ addons/im_livechat/static/src/embed/core/thread_service_patch.js
@@ -4,7 +4,6 @@ import { SESSION_STATE } from "@im_livechat/embed/core/livechat_service";
 
 import { ThreadService, threadService } from "@mail/core/common/thread_service";
 import { prettifyMessageContent } from "@mail/utils/common/format";
-import { createLocalId, onChange } from "@mail/utils/common/misc";
 
 import { markup } from "@odoo/owl";
 
@@ -108,56 +107,6 @@ patch(ThreadService.prototype, {
         if (this.chatbotService.active) {
             this.chatbotService.start();
         }
-    },
-
-    insert(data) {
-        const isUnknown = !(createLocalId(data.model, data.id) in this.store.Thread.records);
-        const thread = super.insert(...arguments);
-        if (thread.type === "livechat" && isUnknown) {
-            if (
-                this.livechatService.displayWelcomeMessage &&
-                !this.chatbotService.isChatbotThread(thread)
-            ) {
-                this.livechatService.welcomeMessage = this.store.Message.insert({
-                    id: this.messageService.getNextTemporaryId(),
-                    body: this.livechatService.options.default_message,
-                    res_id: thread.id,
-                    model: thread.model,
-                    author: thread.operator,
-                });
-            }
-            if (this.chatbotService.isChatbotThread(thread)) {
-                this.chatbotService.typingMessage = this.store.Message.insert({
-                    id: this.messageService.getNextTemporaryId(),
-                    res_id: thread.id,
-                    model: thread.model,
-                    author: thread.operator,
-                });
-            }
-            onChange(thread, "state", () => {
-                if (
-                    ![SESSION_STATE.CLOSED, SESSION_STATE.NONE].includes(this.livechatService.state)
-                ) {
-                    this.livechatService.updateSession({ state: thread.state });
-                }
-            });
-            onChange(thread, "seen_message_id", () => {
-                if (
-                    ![SESSION_STATE.CLOSED, SESSION_STATE.NONE].includes(this.livechatService.state)
-                ) {
-                    this.livechatService.updateSession({ seen_message_id: thread.seen_message_id });
-                }
-            });
-            onChange(thread, "message_unread_counter", () => {
-                if (
-                    ![SESSION_STATE.CLOSED, SESSION_STATE.NONE].includes(this.livechatService.state)
-                ) {
-                    this.livechatService.updateSession({ channel: thread.channel });
-                }
-            });
-            this.store.livechatThread = thread;
-        }
-        return thread;
     },
 
     async update(thread, data) {
