+++ addons/mail/static/src/core/common/chat_window_model.js
@@ -1,12 +1,52 @@
 /* @odoo-module */
 
-import { Record } from "@mail/core/common/record";
+import { Record, modelRegistry } from "@mail/core/common/record";
+import { assignDefined } from "@mail/utils/common/misc";
 
 import { _t } from "@web/core/l10n/translation";
 
 /** @typedef {{ thread?: import("@mail/core/common/thread_model").Thread, folded?: boolean, replaceNewMessageChatWindow?: boolean }} ChatWindowData */
 
 export class ChatWindow extends Record {
+    /** @type {ChatWindow[]} */
+    static records = [];
+    /**
+     * @param {ChatWindowData} [data]
+     * @returns {ChatWindow}
+     */
+    static insert(data = {}) {
+        const chatWindow = this.records.find((c) => c.threadLocalId === data.thread?.localId);
+        if (!chatWindow) {
+            const chatWindow = new ChatWindow(this.store, data);
+            assignDefined(chatWindow, data);
+            let index;
+            const visible = this.env.services["mail.chat_window"].visible;
+            const maxVisible = this.env.services["mail.chat_window"].maxVisible;
+            if (!data.replaceNewMessageChatWindow) {
+                if (maxVisible <= this.records.length) {
+                    const swaped = visible[visible.length - 1];
+                    index = visible.length - 1;
+                    this.env.services["mail.chat_window"].hide(swaped);
+                } else {
+                    index = this.records.length;
+                }
+            } else {
+                const newMessageChatWindowIndex = this.records.findIndex((cw) => !cw.thread);
+                index =
+                    newMessageChatWindowIndex !== -1
+                        ? newMessageChatWindowIndex
+                        : this.records.length;
+            }
+            this.records.splice(index, data.replaceNewMessageChatWindow ? 1 : 0, chatWindow);
+            return this.records[index]; // return reactive version
+        }
+        if (chatWindow.hidden) {
+            this.env.services["mail.chat_window"].makeVisible(chatWindow);
+        }
+        assignDefined(chatWindow, data);
+        return chatWindow;
+    }
+
     /** @type {import("@mail/core/common/store_service").Store} */
     _store;
 
@@ -45,3 +85,5 @@ export class ChatWindow extends Record {
         return !this.folded && !this.hidden;
     }
 }
+
+modelRegistry.add(ChatWindow.name, ChatWindow);
