+++ addons/mail/static/src/core/common/store_service.js
@@ -7,119 +7,33 @@ import { reactive } from "@odoo/owl";
 import { _t } from "@web/core/l10n/translation";
 import { registry } from "@web/core/registry";
 import { debounce } from "@web/core/utils/timing";
+import { modelRegistry } from "./record";
 
 export class Store {
-    Activity = {
-        /** @type {Object.<number, import("@mail/core/web/activity_model").Activity>} */
-        records: {},
-        /**
-         * @param {import("@mail/core/web/activity_model").Data} data
-         * @param {Object} [param1]
-         * @param {boolean} param1.broadcast
-         * @returns {import("@mail/core/web/activity_model").Activity}
-         */
-        insert: (data, { broadcast = true } = {}) =>
-            this.env.services["mail.activity"].insert(data, { broadcast }),
-    };
-    Attachment = {
-        /** @type {Object.<number, import("@mail/core/common/attachment_model").Attachment>} */
-        records: {},
-        /**
-         * @param {Object} data
-         * @returns {import("@mail/core/common/attachment_model").Attachment}
-         */
-        insert: (data) => this.env.services["mail.attachment"].insert(data),
-    };
-    CannedResponse = {
-        /** @type {Object.<number, import("@mail/core/common/canned_response_model").CannedResponse>} */
-        records: {},
-        /**
-         * @param {Object} data
-         * @returns {import("@mail/core/common/canned_response_model").CannedResponse}
-         */
-        insert: (data) => this.env.services["mail.messaging"].insertCannedResponse(data),
-    };
-    ChannelMember = {
-        /** @type {Object.<number, import("@mail/core/common/channel_member_model").ChannelMember>} */
-        records: {},
-        /**
-         * @param {Object|Array} data
-         * @returns {import("@mail/core/common/channel_member_model").ChannelMember}
-         */
-        insert: (data) => this.env.services["discuss.channel.member"].insert(data),
-    };
-    ChatWindow = {
-        /** @type {import("@mail/core/common/chat_window_model").ChatWindow[]} */
-        records: [],
-        /**
-         * @param {ChatWindowData} [data]
-         * @returns {import("@mail/core/common/chat_window_model").ChatWindow}
-         */
-        insert: (data = {}) => this.env.services["mail.chat_window"].insert(data),
-    };
-    Follower = {
-        /** @type {Object.<number, import("@mail/core/common/follower_model").Follower>} */
-        records: {},
-        /**
-         * @param {import("@mail/core/common/follower_model").Data} data
-         * @returns {import("@mail/core/common/follower_model").Follower}
-         */
-        insert: (data) => this.env.services["mail.thread"].insertFollower(data),
-    };
-    Message = {
-        /** @type {Object.<number, import("@mail/core/common/message_model").Message>} */
-        records: {},
-        /**
-         * @param {Object} data
-         * @returns {import("@mail/core/common/message_model").Message}
-         */
-        insert: (data) => this.env.services["mail.message"].insert(data),
-    };
-    Notification = {
-        /** @type {Object.<number, import("@mail/core/common/notification_model").Notification>} */
-        records: {},
-        /**
-         * @param {Object} data
-         * @returns {import("@mail/core/common/notification_model").Notification}
-         */
-        insert: (data) => this.env.services["mail.message"].insertNotification(data),
-    };
-    NotificationGroup = {
-        /** @type {import("@mail/core/common/notification_group_model").NotificationGroup[]} */
-        records: [],
-        /**
-         * @param {Object} data
-         * @returns {import("@mail/core/common/notification_group_model").NotificationGroup}
-         */
-        insert: (data) => this.env.services["mail.message"].insertNotificationGroups(data),
-    };
-    Persona = {
-        /** @type {Object.<number, import("@mail/core/common/persona_model").Persona>} */
-        records: {},
-        /**
-         * @param {import("@mail/core/common/persona_model").Data} data
-         * @returns {import("@mail/core/common/persona_model").Persona}
-         */
-        insert: (data) => this.env.services["mail.persona"].insert(data),
-    };
-    RtcSession = {
-        /** @type {Object.<number, import("@mail/discuss/call/common/rtc_session_model").RtcSession>} */
-        records: {},
-        /**
-         * @param {Object} data
-         * @returns {number, import("@mail/discuss/call/common/rtc_session_model").RtcSession}
-         */
-        insert: (data) => this.env.services["discuss.rtc"].insertSession(data),
-    };
-    Thread = {
-        /** @type {Object.<string, import("@mail/core/common/thread_model").Thread>} */
-        records: {},
-        /**
-         * @param {Object} data
-         * @returns {import("@mail/core/common/thread_model").Thread}
-         */
-        insert: (data) => this.env.services["mail.thread"].insert(data),
-    };
+    /** @type {typeof import("@mail/core/web/activity_model").Activity} */
+    Activity;
+    /** @type {typeof import("@mail/core/common/attachment_model").Attachment} */
+    Attachment;
+    /** @type {typeof import("@mail/core/common/canned_response_model").CannedResponse} */
+    CannedResponse;
+    /** @type {typeof import("@mail/core/common/channel_member_model").ChannelMember} */
+    ChannelMember;
+    /** @type {typeof import("@mail/core/common/chat_window_model").ChatWindow} */
+    ChatWindow;
+    /** @type {typeof import("@mail/core/common/follower_model").Follower} */
+    Follower;
+    /** @type {typeof import("@mail/core/common/message_model").Message} */
+    Message;
+    /** @type {typeof import("@mail/core/common/notification_model").Notification} */
+    Notification;
+    /** @type {typeof import("@mail/core/common/notification_group_model").NotificationGroup} */
+    NotificationGroup;
+    /** @type {typeof import("@mail/core/common/persona_model").Persona} */
+    Persona;
+    /** @type {typeof import("@mail/discuss/call/common/rtc_session_model").RtcSession} */
+    RtcSession;
+    /** @type {typeof import("@mail/core/common/thread_model").Thread} */
+    Thread;
 
     /**
      * @param {import("@web/env").OdooEnv} env
@@ -257,6 +171,19 @@ export const storeService = {
      */
     start(env, services) {
         const res = reactive(new Store(env, services));
+        for (const [name, Model] of modelRegistry.getEntries()) {
+            if (res[name]) {
+                throw new Error(
+                    `There must be no duplicated Model Names (duplicate found: ${name})`
+                );
+            }
+            // classes cannot be made reactive because they are functions and they are not supported.
+            // work-around: make an object whose prototype is the class, so that static props become
+            // instance props.
+            const entry = Object.assign(Object.create(Model), { env, store: res });
+            entry.records = JSON.parse(JSON.stringify(Model.records));
+            res[name] = entry;
+        }
         onChange(res.Thread, "records", () => res.updateBusSubscription());
         services.ui.bus.addEventListener("resize", () => {
             if (!services.ui.isSmall) {
